ADEN-9924 Split video handler into bfaa and bfab handlers.
